Move detail methods to detail VM
